requirements.txt
alembic==1.10.3
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.6
pycparser==2.21
pyOpenSSL==23.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
Werkzeug==2.2.3

attrs~=23.1.0
requests~=2.29.0
celery~=5.2.7
kombu~=5.2.4